Added interfaces and updated to use HttpClient 4.3
Updated the library to use HttpClient 4.3
Using a proxy with the project was difficult if not impossible when
using ClientConnectionManager. It is also deprecated. Instead, allow
direct supply of a pre-built HttpClient.
A few changes were needed to be made to support the new HttpClient.

Also make DeltaSyncClient*.java implement interface versions of the
classes, to potentially allow dynamic proxy classes (such as Spring's
AOP)